# perf(weave): download dataset with images in rows in parallel