--- a/weave/trace/vals.py
+++ b/weave/trace/vals.py
@@ -439,6 +439,12 @@ class WeaveTable(Traceable):
         if self.table_ref is None:
             return
 
+        wc = get_weave_client()
+        if not wc:
+            raise ValueError(
+                "No weave client found. Weave client required to iterate over remote tables"
+            )
+
         page_index = 0
         page_size = 100
         while True:
@@ -474,6 +480,8 @@ class WeaveTable(Traceable):
                     logger.debug(msg)
                     self._prefetched_rows = None
 
+            # Process rows in parallel using the weave client's future executor
+            futures = []
             for i, item in enumerate(response.rows):
                 new_ref = self.ref.with_item(item.digest) if self.ref else None
                 # Here, we use the raw rows if they exist, otherwise we use the
@@ -487,9 +495,25 @@ class WeaveTable(Traceable):
                     if self._prefetched_rows is None
                     else self._prefetched_rows[page_index * page_size + i]
                 )
-                res = from_json(val, self.table_ref.project_id, self.server)
-                res = make_trace_obj(res, new_ref, self.server, self.root)
-                yield res
+
+                def process_row(val: Any, new_ref: RefWithExtra) -> Any:
+                    if not self.table_ref:
+                        raise ValueError("No table ref found")
+                    return make_trace_obj(
+                        from_json(val, self.table_ref.project_id, self.server),
+                        new_ref,
+                        self.server,
+                        self.root,
+                    )
+
+                future = wc.future_executor.defer(
+                    lambda v=val, r=new_ref: process_row(v, r)
+                )
+                futures.append(future)
+
+            # Yield results as they complete
+            for future in futures:
+                yield future.result()
 
             if len(response.rows) < page_size:
                 break