--- a/weave/trace/vals.py
+++ b/weave/trace/vals.py
@@ -12,7 +12,10 @@ from pydantic import v1 as pydantic_v1
 
 from weave.trace import box
 from weave.trace.context.tests_context import get_raise_on_captured_errors
-from weave.trace.context.weave_client_context import get_weave_client
+from weave.trace.context.weave_client_context import (
+    get_weave_client,
+    require_weave_client,
+)
 from weave.trace.object_record import ObjectRecord
 from weave.trace.op import is_op, maybe_bind_method
 from weave.trace.refs import (
@@ -439,11 +442,7 @@ class WeaveTable(Traceable):
         if self.table_ref is None:
             return
 
-        wc = get_weave_client()
-        if not wc:
-            raise ValueError(
-                "No weave client found. Weave client required to iterate over remote tables"
-            )
+        wc = require_weave_client()
 
         page_index = 0
         page_size = 100