## Description

<!--
Include reference to internal ticket "Fixes WB-NNNNN" and/or GitHub issue "Fixes #NNNN" (if applicable)
-->

[WB-23955](https://wandb.atlassian.net/browse/WB-23955)

Use the future executor to parallelize row requests. The only issue I can see is when the queue is full, but I think in that case this executes exactly as fast as it currently does. 

This pr: 
- utilizes the async_batch_processor to handle row download requests, massively speeding up row downloads when rows have objects in them (like images). 

## Testing

| Before | After |
| -------- | ------- |
| `29.7s` | `2.24s` | 
| <img width="818" alt="Screenshot 2025-03-24 at 10 13 51 AM" src="https://github.com/user-attachments/assets/9496b7fd-30b5-4e30-8523-d2346f2c5a3c" /> | <img width="833" alt="Screenshot 2025-03-24 at 10 09 59 AM" src="https://github.com/user-attachments/assets/677afb98-287f-462e-81cd-9cd6534b71bf" /> | 


Image dataset test (run eval against downloaded dataset):
| size | prod eval time (s) | branch  eval time (s) | 
| -------- | ------- | ------- |
| 10 | `1.92s` | `0.66s` |
| 100 | `14.66s` | `0.96s` |
| 1000 | `175.16s` | `20.16s` |


[WB-23955]: https://wandb.atlassian.net/browse/WB-23955?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ